+++ src/java/class-sample/CE3104-POO-HernandezValerie/Geometry/src/ce3104/poo/geometry/shapes/Trapezoid.java
@@ -2,20 +2,20 @@ package ce3104.poo.geometry.shapes;
 
 public final class Trapezoid extends Shape {
 
-    private final Float side;
-    private final Float smallBase;
-    private final Float largeBase;
-    private final Float height;
+    private final Integer side;
+    private final Integer smallBase;
+    private final Integer largeBase;
+    private final Integer height;
 
-    public Trapezoid(Float side, Float smallBase, Float largeBase, Float height) {
+    public Trapezoid(Integer side, Integer smallBase, Integer largeBase, Integer height) {
         this.side = side;
         this.smallBase = smallBase;
         this.largeBase = largeBase;
         this.height = height;
     }
 
-    public Trapezoid(Float smallBase, Float largeBase, Float height) {
-        this.side = (float) (Math.sqrt(Math.pow(height, 2) + Math.pow(((largeBase - smallBase) / 2), 2)));
+    public Trapezoid(Integer smallBase, Integer largeBase, Integer height) {
+        this.side = (int) (Math.sqrt(Math.pow(height, 2) + Math.pow(((largeBase - smallBase) / 2), 2)));
         this.smallBase = smallBase;
         this.largeBase = largeBase;
         this.height = height;
@@ -23,11 +23,11 @@ public final class Trapezoid extends Shape {
 
     @Override
     public Float getPerimeter() {
-        return ((2 * this.side) + this.smallBase + this.largeBase);
+        return (float) ((2 * this.side) + this.smallBase + this.largeBase);
     }
 
     @Override
     public Float getArea() {
-        return ((this.height * (this.smallBase + this.largeBase)) / 2);
+        return (float) ((this.height * (this.smallBase + this.largeBase)) / 2);
     }
 }
